+++ lexer/error.ts
@@ -4,7 +4,7 @@ import type { Token } from './lexer';
  * Custom error class so that we can display the already-extracted tokens
  * which will help the user see where the lexer is up to and got stuck
  */
-export default class LexerError extends TypeError {
+export default class Error extends TypeError {
 	private tokens;
 
 	constructor (message: string, tokens: Token[]) {
