+++ lexer/lexer.spec.ts
