+++ tests/fixtures/lexer_sample.out
